refactor: C3. Break Cyclic Dependency in Producer Package
- Create new BearerTokenResponseFactory utility class
- Extract all HTTP response creation logic from BearerTokenStatus enum
- Remove createResponse() methods from BearerTokenStatus to break cycle
- Update BearerTokenResult to delegate to the factory
- Eliminate bidirectional dependency between BearerTokenResult and BearerTokenStatus
- Update sonar.adoc to mark task C3 as completed (1/6 tasks done)

This refactoring improves separation of concerns, enhances testability,
and completely eliminates the cyclic dependency identified by SonarCloud.
